third_party/blink/web_tests/external/wpt/compute-pressure/compute_pressure_different_quantizations_across_iframes.tentative.https.window-expected.txt

This is a testharness.js-based test.
FAIL PressureObserver with a new quantization schema stops all other active observers assert_equals: observer2.observe() should have stopped observer1; the two observers have different quantization schemes expected 0 but got 1
Harness: the test ran to completion.
